+++ client/api/types/equipment.ts
@@ -1,7 +1,5 @@
 import { ApiModel } from '@/api'
 
-export type Request = object
-
 export interface Response {
     count?: number
     items: ApiModel.Equipment[]