+++ client/api/types/weather.ts
@@ -1,6 +1,6 @@
 import { ApiModel } from '@/api'
 
-/* Statistic */
+/* Weather */
 export interface RequestHistory {
     start_date: string
     end_date: string